services/user-service/requirements.txt
# Pinned to older versions — known CVEs for Trivy to find
Flask==2.0.1
PyJWT==2.1.0
psycopg2-binary==2.9.1
Werkzeug==2.0.1
